beginning to figure out the cmd submission process
Found where the interesting bit of the cmdstream is hiding, after
the first 5 words beyond the contex restore cmd.  The first 5
words appears to be some common pre-emble packet?  Figured out the
length field (the previous packet has the length of the next, and
the address of the next if it is a branch).

The command submission appears to branch between entries in the
ringbuffer memory mapped to the gpu (where markers are also inserted)
and the gpu mapped buffers passed in to issueibcmds ioctl.  In
z180_cmdstream_issueibcmds(), the kernel patches up the address to
the next command back into the ringbuffer (for which userspace does
not have the address).