redump: begin adding redump tool
Some helper functions to generate a .rd file (basically just simple
type/length/value type format), logging the test info, commands,
gpu addresses, context and cmdstream buffers.  And the corresponding
redump utility that reads in the .rd files from a serious of similar
operations with varying parameters and dumps out a side-by-side
comparison of the buffer contents (currently in html table form)
with gpuaddrs, and similar parts highlighted to make it easier to
see where the different parameters end up in the cmdstream.