add test to render to fb
Add test to attempt to render to fb..  seems this doesn't map the
fb properly, so it is really non-functional.  But I think I figured
out how to do that via msm-drm.c API, so this is probably not needed
anymore.  But figured I should add it incase it because useful to
resurrect later.