organize directory tree a bit
Just some house-keeping.. having everything in the root directory
will get increasingly confusing when I start adding 3d tests.

Note: just a VPATH hack in the makefile for now to avoid changing
it.. not wanting to muck with makefiles is the reason I started
with a flat directory structure in the first place.