pgmdump: add CF instruction opc
This looks like it could be the op-code.. EXEC/EXEC_END matches up to
the screenshot in optimize-for-adreno.pdf, and seems to make sense for
the other shaders.  And this filters out the CF instructions with non-
sensical addr/cnt for shaders with loops (where I'd expect flow cntrl
instructions to be).