pgmdump: further refine CF parsing
It appears that the EXEC type CF instruction differentiates between
ALU and FETCH instructions in it's body.  This also seems to match
the SERIALIZE() bit in optimize-for-adreno.pdf, although that name
doesn't seem to make sense.  But the low bit in each two bit pair
predicts whether a 96bit instruction in the payload should be
interpreted as a FETCH or ALU instruction.

Also a few other cleanups and updated bitfield description for CF
instructions.