pgmdump: more CF updates
Change a bit how CF instructions are printed, to more or less match
the screenshot in optimize-for-adreno.pdf.  Also, add "ALLOC PARAM"
matching what is in the screenshot.  This seems to allocate space
for exporting varyings (vs) or pixels (fs), the SIZE() value is the
number of exported values minus one.  For vertex shaders, when
gl_Position/gl_PointSize is exported, a different CF instruction
seems to be used.. calling this "ALLOC COORD".