pgmdump: change src3 swizzle
I've found some MULADDv's which need a swizzle on src3.. and if I
interpret src3 swizzle/etc in the same way as src1 and src2, this
fits better.  It still leaves a question mark about scalar ops that
should take two srcs, although that was already an open question.
It does seem to fit for single src scalar ops, and it would make
sense that instruction dispatch has 3 ports to fetch register vals
in the same way, and then use them for either a 3 src vector op or
a 2 src vector op plus a scalar op.  There is still a possibility
that there is a different interpretation for 2 src scalar ops, but
it seems like the bitfields are layed out for easy parsing (which
is also true for r600 which at least seems to have some similar
concepts, so I guess we can assume that it originated from some of
the same minds).