pgmdump: add constant 0.0/1.0 swiz values
For fetch instructions, the dst swizzle can specify to use the
constant 0.0 or 1.0.. this seems to be used for widening, for
example loading 3 element attribute into a vec4.  The r600 does
something like this (see DST_SEL_{X,Y,Z,W} in VTX_DWORD1_GPR,
for example), and it seems to fit what the patched up VERTEX
fetch instructions look like.