pgmdump: update VERTEX fetch parsing
I don't know if these are really the right names, but in the patched
shader in the command stream (after resolving attributes), there are
two parameters that seem to combine to give you the index to the
attribute data.  If you re-order the fetches in order of increasing
memory location of attribute data, what I'm calling the swizzle seems
to follow a pattern of 0, 1, 2, 0, 1, 2, ... (or x, y, z, x, y, z),
and when it rolls over the CONST() value increases.  I don't really
know at this point if R0.xyz is containing three addresses, or this
value is something other than src swizzle/select.  Although I'm
getting close to the point where I can write my own shader to save
off the value in R0 and find out.