fdre: add regdump test
An attempt to dump R0 register contents to see what is input to the
VERTEX fetch instruction.  And probably a useful playground for other
such experiments.

Not quite sure of what to make of the results yet, but it does seem
that COLORX_32_32_32_32_FLOAT is the correct format for exporting
register values directly.  The order in memory ends up being zyxw.