fdre: begin rearranging stuff for multi-tile support
Re-arrange to only do gmem2mem on flush and a few other related cleanups.
For multi-tile drawing, we'll want to do clear plus all the draw calls
in one go for each tile before moving on to the next tile, only xfering
back out to main memory at the end of the tile.

What I think we can do is setup all the draw cmds in one ringbuffer,
and then at flush time, use a 2nd ringbuffer to setup scissor, viewport,
etc and then IB to the draw draw cmds, using the same draw cmds for each
tile.  Although this means wraparound and flushing at arbitrary points
won't really work transparently.  For now use waittimestamp to wait for
the GPU to finish, and then reset the ringbuffer position back to zero.
We probably want to have multiple pairs of buffers that we cycle through
to avoid stalling.