fdre: get consts/varyings/etc from assembler
Now we figure out the position/order/idx of uniforms/attributes/
textures from the assembler.  And consts come from the shader, so now
they don't have to be hacked in as uniforms.  This is needed in order
to deal properly with more complex fragment shaders, because we need
to know which consts/uniforms belong to the vertex shader and which
to the fragment shader.