cltool: add way to poke in our own shader to disasm
A bit fugly, because we have to find the location of the shader binary
from the ocl compiler in memory, in order to overwrite it with our own.
But seems to do the job.

Also, add missing crtbegin/end stuff needed to make cmdline args work.
This was missed on initial checkin because normally I have .gitignore
ignore .o files ;-)