cffdump: use envytools
Ripped out most of the register parsing, other than a few special cases,
and use envytools instead.  At the moment it is somewhat less featureful
than the old code for a2xx, until the a2xx.xml file is updated with all
the register bitfields.  It should be equivalent for a3xx, where most of
the bitfield parsing had not yet been added.