fdre -> fdre-a2xx
Make room for fdre-a3xx..  probably not the most elegant way, but I
think it is easier just to have two copies of fdre for a2xx vs a3xx.
Even the test programs would have to be slightly different due to the
new shader ISA (and therefore new assembly syntax).  But it is better to
not spend too much time on code sharing, and rather just figure out what
we need to figure out in order to start the gallium driver work.