fdre-a3xx: emit shader state and add simple test
Emit shader state, and add enough parts to (in theory) do a glClear()
type operation to GMEM.. although I can't actually verify the results
until I implement gmem2mem to get the results back out to system memory.
But it is no longer hanging on CP_LOAD_STATE and CP_DRAW_INDX, so it
seems that the shaders should be running and doing *something*
successfully.