fdre-a3xx: add @buf support
Shaders can use the @buf attribute to declare buffer parameters (which
can be, for example, used as compute input and/or output buffers).

Storage can be bound to the buf via fd_set_buf().  At runtime the
specified const will be loaded with the actual buffer address.  For
example, in shader:

   @buf(c5.x) inbuf
   @buf(c5.z) outbuf
   (sy)(ss)mov.s32s32 r0.w, 0
   mov.f32f32 r1.y, c5.z
   mov.f32f32 r1.z, c5.x
   (rpt5)nop
   ldg.u32 r1.z,g[r1.z], 1
   (sy)stg.u32 g[r1.y],r1.z, 1
   end

And then in program:

   fd_set_buf(state, "inbuf", inbuf);
   fd_set_buf(state, "outbuf", outbuf);