cffdump: spiff it out a bit more
In particular, add --draw option to dump individual draws.  In some
cases, when we have binning, it would be nice to be able to specify even
to dump a particular draw for a particular tile (and some awareness of
how the tiling workings wrt. IB's).  But for simple stuff this is enough