cffdump: print out buffer info on gpuaddr's
Sometimes it is useful to be able to see when something is a pointer not
to the start of a buffer, or size of buffer.

Currently a5xx only, since otherwise we don't have a good way to know
when a register is a gpuaddr.  Maybe we could look for _BASE and _ADDR
and so on?